Remove header `meevax/chrono/duration.hpp`
